feat: changed input system for more fluidity (still need to make clocks)

+++ engine/include/event/MoveEvent.hpp
@@ -12,6 +12,7 @@
 
 #include <netinet/in.h>
 #include <vector>
+#include <SFML/Graphics.hpp>
 
 namespace potEngine
 {
@@ -62,7 +63,7 @@ namespace potEngine
 
             auto position = _entity->getComponent<PositionComponent>()->get()->_position;
             auto username = _entity->getComponent<PlayerComponent>()->get()->username;
-            int speed = _entity->getComponent<MovementComponent>()->get()->speed;
+            float speed = _entity->getComponent<MovementComponent>()->get()->speed;
             if (info->event == MOVE_UP && position[1] > 0)
                 position[1] = (position[1] > speed) ? position[1] - speed : 0;
             if (info->event == MOVE_DOWN && position[1] < 1080)
